refactor(mcp): collapse sub-threshold product MCP files
Merge dispatcher/request/response/errors into server.rs; merge
selection/injection into product_catalog.rs; drop ProductMcpLaunchSelector
and ProductMcpCapabilityTokenMinter blanket-impl traits in favour of
stored closures; amend all spec guides that named the deleted files.

Arc<dyn> trait simplification folded in (not deferred) — same swarm-generated
ceremony in the same module, S effort, near-zero blast radius.

Out-of-scope and deferred: assembly.rs (475 lines) and product_registry.rs
(455 lines) approaching split thresholds.
