+++ anyharness/crates/anyharness-lib/src/domains/sessions/mcp_bindings/mod.rs
@@ -2,10 +2,8 @@ pub mod acp;
 pub mod assembly;
 pub mod contract;
 pub mod crypto;
-pub mod injection;
 pub mod model;
 pub mod product_catalog;
 pub mod product_launch;
 pub mod product_registry;
-pub mod selection;
 pub mod summaries;
