+++ anyharness/crates/anyharness-lib/src/domains/sessions/mcp_bindings/product_launch.rs
@@ -12,50 +12,19 @@ pub struct ProductMcpSelectionContext<'a> {
     pub session: &'a SessionRecord,
 }
 
-pub trait ProductMcpLaunchSelector: Send + Sync {
-    fn should_attach(&self, ctx: ProductMcpSelectionContext<'_>) -> anyhow::Result<bool>;
-}
-
-impl<F> ProductMcpLaunchSelector for F
-where
-    F: for<'a> Fn(ProductMcpSelectionContext<'a>) -> anyhow::Result<bool> + Send + Sync,
-{
-    fn should_attach(&self, ctx: ProductMcpSelectionContext<'_>) -> anyhow::Result<bool> {
-        self(ctx)
-    }
-}
-
-pub trait ProductMcpCapabilityTokenMinter: Send + Sync {
-    fn mint_capability_token(&self, workspace_id: &str, session_id: &str)
-        -> anyhow::Result<String>;
-}
-
-impl<F> ProductMcpCapabilityTokenMinter for F
-where
-    F: Fn(&str, &str) -> anyhow::Result<String> + Send + Sync,
-{
-    fn mint_capability_token(
-        &self,
-        workspace_id: &str,
-        session_id: &str,
-    ) -> anyhow::Result<String> {
-        self(workspace_id, session_id)
-    }
-}
-
 #[derive(Clone)]
 pub struct ProductMcpLaunchRegistration {
     definition: &'static ProductMcpDefinition,
-    selector: Arc<dyn ProductMcpLaunchSelector>,
-    token_minter: Arc<dyn ProductMcpCapabilityTokenMinter>,
+    selector: Arc<dyn for<'a> Fn(ProductMcpSelectionContext<'a>) -> anyhow::Result<bool> + Send + Sync>,
+    token_minter: Arc<dyn Fn(&str, &str) -> anyhow::Result<String> + Send + Sync>,
     launch_extras: SessionLaunchExtras,
 }
 
 impl ProductMcpLaunchRegistration {
     pub fn new(
         definition: &'static ProductMcpDefinition,
-        selector: Arc<dyn ProductMcpLaunchSelector>,
-        token_minter: Arc<dyn ProductMcpCapabilityTokenMinter>,
+        selector: Arc<dyn for<'a> Fn(ProductMcpSelectionContext<'a>) -> anyhow::Result<bool> + Send + Sync>,
+        token_minter: Arc<dyn Fn(&str, &str) -> anyhow::Result<String> + Send + Sync>,
     ) -> Self {
         Self {
             definition,
@@ -91,7 +60,7 @@ impl ProductMcpLaunchRegistration {
     }
 
     pub fn should_attach(&self, ctx: ProductMcpSelectionContext<'_>) -> anyhow::Result<bool> {
-        self.selector.should_attach(ctx)
+        (self.selector)(ctx)
     }
 
     pub fn mint_capability_token(
@@ -99,7 +68,6 @@ impl ProductMcpLaunchRegistration {
         workspace_id: &str,
         session_id: &str,
     ) -> anyhow::Result<String> {
-        self.token_minter
-            .mint_capability_token(workspace_id, session_id)
+        (self.token_minter)(workspace_id, session_id)
     }
 }
